Fixed bugs on edges
Two bugs: First bug was when running something like '--+', the compressed graph would have edge with label: 'memory[data_pointer] += -1', which is silly, so I changed it to correctly display: '... -= 1'. Second bug was relating to having any non-instruction symbols in the code, which would lead to node numbers getting large, which is not only weird numbering-wise, but it also lead to edges not connecting to nodes if non-instructions seperated them. Now fixed :)